backend/requirements.txt
elevenlabs
openai
python-dotenv
pydantic
PyJWT
python-multipart
elevenlabs
networkx
chromadb
tiktoken
python-multipart
